Refactor DesignCanvas into sub-components and custom hook
Extracted view rendering logic from DesignCanvas.jsx into specialized
components (GridRenderer, ShapeRenderer, HandleRenderer) inside the
new canvas/ directory. Migrated event handling and state management
for canvas interactions into a new custom hook (useCanvasInteraction).
Created a comprehensive ANALYSIS_AND_PLAN.md guide for contributors.
