+++ frontend/npm_output.log
@@ -1,0 +1,9 @@
+
+> frontend@0.0.0 dev
+> vite
+
+
+  VITE v3.2.11  ready in 472 ms
+
+  ➜  Local:   http://localhost:5173/
+  ➜  Network: use --host to expose
